+++ src/CsvInterface.php
@@ -16,4 +16,20 @@ interface CsvInterface extends StreamInterface
      * @return null|bool|array NULL if an invalid handle is supplied or FALSE on other errors, including end of file.
      */
     public function getCsv($length, $delimiter, $enclosure, $escape);
+
+    /**
+     * Set an option metadata configured in this resource
+     *
+     * @param string $key The metadata key
+     * @param mixed $value The value
+     * @return void
+     */
+    public function setMetadata(string $key, $value);
+
+    /**
+     * Returns the underlying resource
+     *
+     * @return mixed The underlying resource
+     */
+    public function getResource();
 }