+++ tests/splfile/TraitResource.php
@@ -6,7 +6,9 @@ trait TraitResource
     protected function getResource($file, $debug = false)
     {
         $file = $this->dir . '/' . $file;
-        $this->parser->setOption('debug', $debug);
+        if (isset($this->parser)) {
+            $this->parser->setOption('debug', $debug);
+        }
         return new \voilab\csv\CsvSplFile(new \SplFileObject($file), [ 'debug' => $debug ]);
     }
 }